requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Werkzeug==2.3.7
google-auth==2.22.0
google-auth-oauthlib==1.0.0
python-dotenv==1.0.0
reportlab==4.0.4 